Entire Project Refactored back to MVC, App_Start & Testing; Architecture.
I reverted the entire project to conform to MVC design pattern. This will allow for Team Development, Scalable code and logical organisation.

Please use the testing folder for all tests and testing, i have linked it with a ID within IntelliJ so it knows all unit test are in this folder.

Thanks,
Brad.